// test/lib/api-client/models/GetAdminSmartApps200ResponseInner.ts
export interface GetAdminSmartApps200ResponseInner {
    /**
     * Client ID
     * @type {string}
     * @memberof GetAdminSmartApps200ResponseInner
     */
    id?: string;
    /**
     * Client identifier
     * @type {string}
     * @memberof GetAdminSmartApps200ResponseInner
     */
    clientId?: string;
    /**
     * Application name
     * @type {string}
     * @memberof GetAdminSmartApps200ResponseInner
     */
    name?: string;
    /**
     * Application description
     * @type {string}
     * @memberof GetAdminSmartApps200ResponseInner
     */
    description?: string;
    /**
     * Whether app is enabled
     * @type {boolean}
     * @memberof GetAdminSmartApps200ResponseInner
     */
    enabled?: boolean;
    /**
     * OAuth protocol
     * @type {string}
     * @memberof GetAdminSmartApps200ResponseInner
     */
    protocol?: string;
    /**
     * Whether app is public client
     * @type {boolean}
     * @memberof GetAdminSmartApps200ResponseInner
     */
    publicClient?: boolean;
    /**
     * 
     * @type {Array<string>}
     * @memberof GetAdminSmartApps200ResponseInner
     */
    redirectUris?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof GetAdminSmartApps200ResponseInner
     */
    webOrigins?: Array<string>;
    /**
     * 
     * @type {object}
     * @memberof GetAdminSmartApps200ResponseInner
     */
    attributes?: object;
}